+++ changelog.md
@@ -6,3 +6,4 @@
 - Fixed random accessed from netty thread crash when orbit sounds play. (#473).
 - Fixed The Ti-69s red and green lights being swapped from the item texture (#479).
 - Added painting translations (#485).
+- The Nasa workbench now works with recipes with less than 14 ingredients (#483).